+++ src/main/java/com/example/ordemPedidos/entities/Endereco.java
@@ -18,7 +18,7 @@ public class Endereco implements Serializable{
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	
 	private String logradouro;
 	private String numero;
@@ -38,7 +38,7 @@ public class Endereco implements Serializable{
 	public Endereco() {
 	}
 
-	public Endereco(Long id, String logradouro, String numero, String complemento, String bairro, String cep, Cliente cliente,Cidade cidade) {
+	public Endereco(Integer id, String logradouro, String numero, String complemento, String bairro, String cep, Cliente cliente,Cidade cidade) {
 		super();
 		this.id = id;
 		this.logradouro = logradouro;
@@ -50,11 +50,11 @@ public class Endereco implements Serializable{
 		this.cidade = cidade;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
