+++ src/main/java/com/example/ordemPedidos/entities/PagamentoComBoleto.java
@@ -21,7 +21,7 @@ public class PagamentoComBoleto extends Pagamento{
 	public PagamentoComBoleto() {
 	}
 
-	public PagamentoComBoleto(Long id, EstadoPagamento estado, Pedido pedido, Date dataVencimento, Date dataPagamento) {
+	public PagamentoComBoleto(Integer id, EstadoPagamento estado, Pedido pedido, Date dataVencimento, Date dataPagamento) {
 		super(id, estado, pedido);
 		this.dataVencimento = dataVencimento;
 		this.dataPagamento = dataPagamento;
