+++ src/main/java/com/example/ordemPedidos/entities/PagamentoComCartao.java
@@ -14,7 +14,7 @@ public class PagamentoComCartao extends Pagamento{
 	public PagamentoComCartao() {
 	}
 
-	public PagamentoComCartao(Long id, EstadoPagamento estado, Pedido pedido, Integer numeroDeParcelas) {
+	public PagamentoComCartao(Integer id, EstadoPagamento estado, Pedido pedido, Integer numeroDeParcelas) {
 		super(id, estado, pedido);
 		this.numeroDeParcelas = numeroDeParcelas;
 	}
