+++ src/main/java/com/example/ordemPedidos/entities/Produto.java
@@ -22,7 +22,7 @@ public class Produto implements Serializable{
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	
 	private String nome;
 	
@@ -42,7 +42,7 @@ public class Produto implements Serializable{
 	public Produto() {
 	}
 
-	public Produto(Long id, String nome, Double preco) {
+	public Produto(Integer id, String nome, Double preco) {
 		super();
 		this.id = id;
 		this.nome = nome;
@@ -58,11 +58,11 @@ public class Produto implements Serializable{
 		return lista;
 	}
 	
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
