+++ src/main/java/com/example/ordemPedidos/repositories/ItemPedidoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.ordemPedidos.entities.ItemPedido;
 
-public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Long>{
+public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Integer>{
 
 }
