+++ src/main/java/com/example/ordemPedidos/repositories/PagamentoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.ordemPedidos.entities.Pagamento;
 
-public interface PagamentoRepository extends JpaRepository<Pagamento, Long>{
+public interface PagamentoRepository extends JpaRepository<Pagamento, Integer>{
 
 }
