+++ src/main/java/com/example/ordemPedidos/repositories/PedidoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.ordemPedidos.entities.Pedido;
 
-public interface PedidoRepository extends JpaRepository<Pedido, Long>{
+public interface PedidoRepository extends JpaRepository<Pedido, Integer>{
 
 }
