+++ src/main/java/com/example/ordemPedidos/services/exceptions/ObjectNotFoundException.java
@@ -4,7 +4,7 @@ public class ObjectNotFoundException extends RuntimeException{
 
 	private static final long serialVersionUID = 1L;
 	
-	public ObjectNotFoundException(Long id) {
+	public ObjectNotFoundException(Integer id) {
 		super("Objeto não ecnontrado com id " + id);
 	}
